requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.9.0
google-genai==1.0.0
google-api-python-client==2.140.0
google-auth==2.35.0
python-dateutil==2.9.0
